requirements.txt
torch
torchvision
pillow
huggingface-hub